Rename schema to ddl
