+++ spectron/__init__.py
@@ -2,4 +2,4 @@
 
 __version__ = "0.2.1"
 
-from . import spectrum_schema
+from . import ddl
